Move api implementation to squeril.api
